Rename create container to create instance
