+++ src/classes/Tools/Instances/CreateInstance.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace dhope0000\LXDClient\Tools\Containers;
+namespace dhope0000\LXDClient\Tools\Instances;
 
 use dhope0000\LXDClient\Tools\Hosts\HostsHaveContainer;
 use dhope0000\LXDClient\Tools\Hosts\Images\ImportImageIfNotHave;
 use dhope0000\LXDClient\Model\Client\LxdClient;
 
-class CreateContainer
+class CreateInstance
 {
     public function __construct(
         LxdClient $lxdClient,
